requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pytest>=8.1.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
